adds compilation checks to run scripts

+++ EXAMPLES/homogeneous_halfspace_HEX8_elastic_absorbing_Stacey_5sides/run_this_example_kernel.sh
@@ -26,6 +26,12 @@ echo
 rm -f change_simulation_type.pl
 ln -s ../../utils/change_simulation_type.pl
 
+# checks if executables were compiled and available
+if [ ! -e ../../bin/xspecfem3D ]; then
+  echo "Please compile first all binaries in the root directory, before running this example..."; echo
+  exit 1
+fi
+
 # cleans output files
 mkdir -p OUTPUT_FILES
 rm -rf OUTPUT_FILES/*
